Add the Spandex simulation system testbench, plus bug fix
About the testbench:
Although unit testing is good and all, some bugs are hard or impossible
to test for without considering the larger system. So this testbench
runs a simulation involving an L2 instance and an LLC instance, connects
them both, and then emulates CPU requests and a memory element on either
side. This allows us to send requests to the L2 as the CPU/accelerator
would, and check for consistency by comparing the results of the requests
to a golden memory reference. Basically, this allows us to use fuzzing.

This testbench has been used to identify the issu listed below, and has
been used to validate other behaviors of the caches.

About the bug fix:
 - Fixes an issue where some AMO operations were implemented incorrectly
   in the Spandex L2.

Another known bug discovered with the testbench: the L2 cache does not
look into its own write buffer when checking for a read hit, so it will
sometimes respond with stale data if there has been a write-through
forward to that line.